Import Project Achilles CRISPR gene effect data

Import Project Achilles CRISPR gene effect data

Achilles(release = NULL, rowRanges = TRUE, colData = TRUE)

Arguments

release
character(1) or NULL. DepMap quarterly release ID (e.g. "20Q3"). If left NULL, defaults to latest supported release.

rowRanges
logical(1). Include gene annotations as genomic ranges.

colData
logical(1). Include cell line metadata.

Value

Achilles.

Note

Updated 2020-10-01.

Assays

effect: CERES data with principle components strongly related to known batch effects removed, then shifted and scaled per cell line so the median nonessential KO effect is 0 and the median essential KO effect is -1.

probability: Probability that knocking out the gene has a real depletion effect using gene_effect.

Examples

object <- Achilles()

#> → Importing achilles_gene_effect.csv from DepMap `20Q3` release.

#> URL:
#>   https://ndownloader.figshare.com/files/24613292

#> → Importing achilles_gene_dependency.csv from DepMap `20Q3` release.

#> URL:
#>   https://ndownloader.figshare.com/files/24613298

#> → Importing sample_info.csv from DepMap `20Q3` release.

#> URL:
#>   https://ndownloader.figshare.com/files/24613394

#> ! Dropping 7 retired genes: CDR1_1038, CRIPAK_285464, ELOA3D_100506888, LUZP6_767558, NPBWR1_2831, OCLM_10896, SPHAR_10638.

#> → Making `GRanges` from Ensembl.

#> → Getting `EnsDb` from AnnotationHub 2.20.2 (2020-04-27).

#> ℹ AH83216: Ensembl 101 EnsDb for Homo sapiens.

#> → Making `GRanges` from `EnsDb`.

#>     Organism: Homo sapiens

#>     Genome build: GRCh38

#>     Ensembl release: 101

#> → Defining `broadClass` using: `geneBiotype, geneName, seqnames`.

#> → Arranging by `geneID`.

#> ℹ 67990 genes detected.

print(object)

#> class: Achilles 
#> dim: 18112 789 
#> metadata(5): date release sessionInfo version wd
#> assays(2): effect probability
#> rownames(18112): a1bg_1 a1cf_29974 ... zzef1_23140 zzz3_26009
#> rowData names(8): broadClass description ... geneName seqCoordSystem
#> colnames(789): ach_000004 ach_000005 ... ach_002204 ach_002446
#> colData names(25): achilles_n_replicates age ... subtype
#>   wtsi_master_cell_id