Import Project Achilles CRISPR gene effect data

Import Project Achilles CRISPR gene effect data

Achilles(release = NULL, rowRanges = TRUE, colData = TRUE)

Arguments

release
character(1) or NULL. DepMap quarterly release ID (e.g. "20Q3"). If left NULL, defaults to latest supported release.

rowRanges
logical(1). Include gene annotations as genomic ranges.

colData
logical(1). Include cell line metadata.

Value

Achilles.

Note

Updated 2020-10-07.

Assays

effect: CERES data with principle components strongly related to known batch effects removed, then shifted and scaled per cell line so the median nonessential KO effect is 0 and the median essential KO effect is -1.

probability: Probability that knocking out the gene has a real depletion effect using gene_effect.

Examples

object <- Achilles(rowRanges = FALSE, colData = FALSE)

#> → Importing achilles_gene_effect.csv from DepMap `20Q3` release.

#> → Caching URL at <https://ndownloader.figshare.com/files/24613292>.

#> → Importing 66c861bae58b_24613292 at /Users/mike/Library/Caches/DepMapAnalysis using vroom::`vroom()`.

#> → Importing achilles_gene_dependency.csv from DepMap `20Q3` release.

#> → Caching URL at <https://ndownloader.figshare.com/files/24613298>.

#> → Importing 66c823d542f3_24613298 at /Users/mike/Library/Caches/DepMapAnalysis using vroom::`vroom()`.

#> → Importing achilles_common_essentials.csv from DepMap `20Q3` release.

#> → Caching URL at <https://ndownloader.figshare.com/files/24613283>.

#> → Importing 11a21795b24_24613283 at /Users/mike/Library/Caches/DepMapAnalysis using vroom::`vroom()`.

#> → Importing common_essentials.csv from DepMap `20Q3` release.

#> → Caching URL at <https://ndownloader.figshare.com/files/24613385>.

#> → Importing 11a2464e993d_24613385 at /Users/mike/Library/Caches/DepMapAnalysis using vroom::`vroom()`.

#> → Importing nonessentials.csv from DepMap `20Q3` release.

#> → Caching URL at <https://ndownloader.figshare.com/files/24613388>.

#> → Importing 11a2522e97da_24613388 at /Users/mike/Library/Caches/DepMapAnalysis using vroom::`vroom()`.

#> Error in Achilles(rowRanges = FALSE, colData = FALSE): object 'retired' not found

print(object)

#> Error in print(object): object 'object' not found